% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceR.R
\name{make_tracery}
\alias{make_tracery}
\title{Generate some text}
\usage{
make_tracery(json)
}
\arguments{
\item{json}{some JSON loaded with \code{jsonlite::read_json} or \code{\link{read_tracery}}}
}
\value{
some text
}
\description{
Having loaded some valid tracery JSON, generate some fun text from it.
}
\author{
David L Miller
}
