% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tracery.R
\name{read_tracery}
\alias{read_tracery}
\title{Read in tracery files}
\usage{
read_tracery(file)
}
\arguments{
\item{file}{path to a tracery JSON file}
}
\value{
an object you can pass to \code{\link{make_tracery}}
}
\description{
Read in JSON files in the tracery format. For now this just uses \code{jsonlite::read_json} but will have further validation later on.
}
\author{
David L Miller
}
